Move helper closer to component
